Super-optimización en la carga del mercado

+++ Common/src/main/java/xyz/cereshost/common/market/Trade.java
@@ -1,6 +1,6 @@
 package xyz.cereshost.common.market;
 
-public record Trade(long id, long time, double price, double qty, boolean isBuyerMaker) {
+public record Trade(long id, long time, float price, float qty, boolean isBuyerMaker) {
 
     public double quoteQty(){
         return price * qty;
